Refactored for Interactable -|--> Item
